--- src/types/credentials.ts
@@ -1,4 +1,0 @@
-export interface Credentials {
-  username: string;
-  password: string;
-}
