--- src/types/exam.ts
@@ -1,7 +1,0 @@
-export interface ExamDay {
-  subject: string;
-  amount: string;
-  date: string;
-  time: string;
-  room: string;
-}
