--- src/types/lesson.ts
@@ -1,7 +1,0 @@
-export interface Lesson {
-  subject: string;
-  room: string;
-  dayOfWeek: number;
-  from: string;
-  to: string;
-}
