--- src/types/news.ts
@@ -1,5 +1,0 @@
-export interface News {
-  title: string;
-  date: string;
-  url: string;
-}
