--- src/types/score.ts
@@ -1,6 +1,0 @@
-export interface Score {
-  subject: string;
-  assignment: string;
-  midterm: string;
-  final: string;
-}
