--- src/types/subject.ts
@@ -1,3 +1,0 @@
-export interface Subject {
-  id: string;
-}
