--- src/types/tuition.ts
@@ -1,11 +1,0 @@
-export interface Tuition {
-  credits: string;
-  tuition: string;
-  discount: string;
-  prevDebt: string;
-  paid: string;
-  healthInsurance: string;
-  healthInsuranceDebt: string;
-  healthInsurancePaid: string;
-  debt: string;
-}
