+++ tsconfig.json
@@ -17,6 +17,5 @@
   },
   "include": [
     "src/**/*",
-    "tests/**/*",
   ],
 }
